Have the clone repos shell out to git.

+++ resque_jobs/clone_new_repo.rb
@@ -1,5 +1,6 @@
 require "bundler/setup"
 require "pathological"
+require "timeout"
 require "lib/script_environment"
 require "lib/resque_job_helper"
 
@@ -9,7 +10,7 @@ class CloneNewRepo
 
   # We're willing to spend up to 5 minutes to clone a repo. This can be necessary for giant repos or
   # when cloning over a slow network connection.
-  TIMEOUT = 5 * 60
+  CLONE_TIMEOUT = 5 * 60
 
   def self.perform(repo_name, repo_url)
     setup
@@ -22,25 +23,21 @@ class CloneNewRepo
     end
 
     logger.info "Cloning the repo #{repo_url} into #{repo_path}. If it's large, this may take awhile."
+    error = nil
     begin
-      Grit::Git.with_timeout(CloneNewRepo::TIMEOUT) do
-        # If this clone operation requires you to enter a password, it will show a prompt and eventually
-        # time out.
-        Grit::Git.new(repo_path).clone({}, repo_url, repo_path)
-      end
-
-      # Unfortunately, Grit::Git.new(...).clone() will not throw an exception if the repo_url was invalid.
-      # It will just... do nothing.
-      # TODO(caleb) We should be able to clone an empty repo; distinguish that case from an invalid one.
-      unless File.exists?(repo_path) && (Grit::Git.new(repo_path).has_refs? rescue false)
-        raise "Unable to clone repo. Perhaps the repo at #{repo_url} is invalid, empty, or unreachable?"
+      # Shell out instead of using Grit::Git.new(repo_path).clone({}, repo_url, repo_path) because
+      # Grit doesn't raise an exception when it has trouble cloning a repo, so there's no good error feedback.
+      Timeout::timeout(CLONE_TIMEOUT) do
+        run_shell("cd '#{REPOS_ROOT}' && git clone '#{repo_url}' #{repo_name}")
       end
       logger.info "Finished cloning the repo #{repo_url}."
-    rescue Exception => error
-      logger.error "Error cloning the repo #{repo_url}: #{error.message}"
+    rescue Timeout::Error => error
+      logger.error "Timed out while trying to clone '#{repo_url}'"
+    rescue StandardError => error
+      logger.error "Unable to clone the repo #{repo_url}"
+    ensure
       # Having an empty directory with a bad checkout is not good and can mess up various parts of Barkeep.
-      # Clean it up.
-      FileUtils.rm_rf(repo_path)
+      FileUtils.rm_rf(repo_path) if error
     end
   end
 end
